requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
numpy==2.1.3
opencv-python==4.10.0.84
pillow==11.0.0
requests==2.32.3
urllib3==2.2.3
